UCB_BioE134_GeneDesign/requirements.txt
pytest
gdown
biopython